refactor: 테스트 파일에서 구현한 Lotto 클래스 분리
