+++ src/__tests__/Lotto.test.js
@@ -1,30 +1,5 @@
 import { describe, test, expect } from "vitest";
-
-class Lotto {
-  static #PRICE_PER_ONE = 1000;
-
-  #count;
-
-  constructor(price) {
-    if (typeof price !== "number") {
-      throw new Error("로또 구입 금액으로 숫자를 입력해야 합니다.");
-    }
-
-    if (price < Lotto.#PRICE_PER_ONE) {
-      throw new Error("로또 구입 금액은 1000원이상이어야 합니다.");
-    }
-
-    this.#count = Lotto.#countLotto(price);
-  }
-
-  get count() {
-    return this.#count;
-  }
-
-  static #countLotto(price) {
-    return Math.floor(price / Lotto.#PRICE_PER_ONE);
-  }
-}
+import { Lotto } from "../domain/index.js";
 
 describe("Lotto 클래스 테스트.", () => {
   test("로또를 구입할 때 숫자가 아닌 값을 받으면 오류가 발생한다.", () => {
