+++ src/domain/index.js
@@ -1,0 +1,1 @@
+export { default as Lotto } from "./Lotto.js";
